完成

+++ app/src/main/java/com/julongsoft/measure/entity/MoneyDetialData.java
@@ -10,10 +10,10 @@ public class MoneyDetialData {
      */
 
     private String contractor;
-    private int approval_ivst;
-    private int plan_ivst;
-    private int period_pay_money;
-    private int total_ivst;
+    private double approval_ivst;
+    private double plan_ivst;
+    private double period_pay_money;
+    private double total_ivst;
 
     public String getContractor() {
         return contractor;
@@ -23,15 +23,15 @@ public class MoneyDetialData {
         this.contractor = contractor;
     }
 
-    public int getApproval_ivst() {
+    public double getApproval_ivst() {
         return approval_ivst;
     }
 
-    public void setApproval_ivst(int approval_ivst) {
+    public void setApproval_ivst(double approval_ivst) {
         this.approval_ivst = approval_ivst;
     }
 
-    public int getPlan_ivst() {
+    public double getPlan_ivst() {
         return plan_ivst;
     }
 
@@ -39,7 +39,7 @@ public class MoneyDetialData {
         this.plan_ivst = plan_ivst;
     }
 
-    public int getPeriod_pay_money() {
+    public double getPeriod_pay_money() {
         return period_pay_money;
     }
 
@@ -47,11 +47,11 @@ public class MoneyDetialData {
         this.period_pay_money = period_pay_money;
     }
 
-    public int getTotal_ivst() {
+    public double getTotal_ivst() {
         return total_ivst;
     }
 
-    public void setTotal_ivst(int total_ivst) {
+    public void setTotal_ivst(double total_ivst) {
         this.total_ivst = total_ivst;
     }
 }
